backend/requirements.txt
aiogram==3.24.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
aiofiles==23.2.1
